Crate mc_rpc_core

Starknet RPC API trait and types

Starkware maintains a description of the Starknet API using the openRPC specification. This crate uses jsonrpsee to define such an API in Rust terms.

Structs

BlockHashAndNumber

DeprecatedCairoEntryPoint

Deprecated Cairo entry point (pre Sierra)

DeprecatedContractClass

Deprecated Cairo contract class (pre Sierra)

DeprecatedEntryPointsByType

EntryPointsByType

EventABIEntry

FunctionABIEntry

FunctionCall

Function call information

SierraContractClass

Cairo sierra contract class

SierraEntryPoint

Cairo sierra entry point

StructABIEntry

SyncStatus

TypedParameter

Enums

BlockId

A block hash, number or tag

BlockTag

A tag specifying a dynamic reference to a block

ContractABIEntry

EventABIType

FunctionABIType

RPCContractClass

Starknet contract class

StructABIType

Syncing

Boolean or SyncStatus

Traits

StarknetRpcApiServer

Server trait implementation for the StarknetRpcApi RPC API.

Functions

to_rpc_contract_class

Returns a ContractClassWrapper from a RPCContractClass

Type Definitions

ABI

BlockHash

BlockNumber

Constructor

ContractABI

ContractAddress

ContractClassVersion

DeprecatedConstructor

DeprecatedExternal

DeprecatedL1Handler

EventName

External

FieldElement

FunctionIndex

FunctionName

L1Handler

Members

Offset

ParameterName

ParameterType

Program

Selector

SierraProgram

Size

StructMember

StructName