Trait mc_rpc_core::StarknetRpcApiServer

pub trait StarknetRpcApiServer: Sized + Send + Sync + 'static {

    // Required methods
    fn block_number(&self) -> RpcResult<u64>;
    fn block_hash_and_number(&self) -> RpcResult<BlockHashAndNumber>;
    fn get_block_transaction_count(&self, block_id: BlockId) -> RpcResult<u128>;
    fn get_storage_at(
        &self,
        contract_address: FieldElement,
        key: FieldElement,
        block_id: BlockId
    ) -> RpcResult<FieldElement>;
    fn call(
        &self,
        request: FunctionCall,
        block_id: BlockId
    ) -> RpcResult<Vec<String>>;
    fn get_class_at(
        &self,
        contract_address: FieldElement,
        block_id: BlockId
    ) -> RpcResult<ContractClass>;
    fn get_class_hash_at(
        &self,
        contract_address: FieldElement,
        block_id: BlockId
    ) -> RpcResult<FieldElement>;
    fn syncing<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = RpcResult<SyncStatusType>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_class(
        &self,
        block_id: BlockId,
        class_hash: FieldElement
    ) -> RpcResult<ContractClass>;
    fn get_block_with_tx_hashes(
        &self,
        block_id: BlockId
    ) -> RpcResult<MaybePendingBlockWithTxHashes>;
    fn get_chain_id(&self) -> RpcResult<String>;
    fn add_invoke_transaction<'life0, 'async_trait>(
        &'life0 self,
        invoke_transaction: BroadcastedInvokeTransaction
    ) -> Pin<Box<dyn Future<Output = RpcResult<InvokeTransactionResult>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn add_deploy_account_transaction<'life0, 'async_trait>(
        &'life0 self,
        deploy_account_transaction: BroadcastedDeployAccountTransaction
    ) -> Pin<Box<dyn Future<Output = RpcResult<DeployAccountTransactionResult>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_transaction_by_block_id_and_index(
        &self,
        block_id: BlockId,
        index: usize
    ) -> RpcResult<Transaction>;

    // Provided method
    fn into_rpc(self) -> RpcModule<Self> { ... }

}

Server trait implementation for the StarknetRpcApi RPC API.

Required Methods

fn block_number(&self) -> RpcResult<u64>

Get the most recent accepted block number

fn block_hash_and_number(&self) -> RpcResult<BlockHashAndNumber>

Get the most recent accepted block hash and number

fn get_block_transaction_count(&self, block_id: BlockId) -> RpcResult<u128>

Get the number of transactions in a block given a block id

fn get_storage_at( &self, contract_address: FieldElement, key: FieldElement, block_id: BlockId ) -> RpcResult<FieldElement>

Get the value of the storage at the given address and key, at the given block id

fn call(&self, request: FunctionCall, block_id: BlockId) -> RpcResult<Vec<String>>

Call a contract function at a given block id

fn get_class_at( &self, contract_address: FieldElement, block_id: BlockId ) -> RpcResult<ContractClass>

Get the contract class at a given contract address for a given block id

fn get_class_hash_at( &self, contract_address: FieldElement, block_id: BlockId ) -> RpcResult<FieldElement>

Get the contract class hash in the given block for the contract deployed at the given address

fn syncing<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = RpcResult<SyncStatusType>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get an object about the sync status, or false if the node is not syncing

fn get_class( &self, block_id: BlockId, class_hash: FieldElement ) -> RpcResult<ContractClass>

Get the contract class definition in the given block associated with the given hash

fn get_block_with_tx_hashes( &self, block_id: BlockId ) -> RpcResult<MaybePendingBlockWithTxHashes>

Get block information with transaction hashes given the block id

fn get_chain_id(&self) -> RpcResult<String>

Get the chain id

fn add_invoke_transaction<'life0, 'async_trait>( &'life0 self, invoke_transaction: BroadcastedInvokeTransaction ) -> Pin<Box<dyn Future<Output = RpcResult<InvokeTransactionResult>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Add an Invoke Transaction to invoke a contract function

fn add_deploy_account_transaction<'life0, 'async_trait>( &'life0 self, deploy_account_transaction: BroadcastedDeployAccountTransaction ) -> Pin<Box<dyn Future<Output = RpcResult<DeployAccountTransactionResult>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Add a Deploy Account Transaction

fn get_transaction_by_block_id_and_index( &self, block_id: BlockId, index: usize ) -> RpcResult<Transaction>

Get the details of a transaction by a given block id and index

Provided Methods

fn into_rpc(self) -> RpcModule<Self>

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.