Module mc_rpc_core::types

Structs

ContractData

Holds the data and proofs for a specific contract.

RpcGetProofInput

RpcGetProofOutput

Holds the membership/non-membership of a contract and its associated contract contract if the contract exists.